마이페이지에 있는 세부 내용 컴포넌트로 옮기기
